正确实现工作区诊断，工作区诊断现在不会占用过多性能
